Additional updates to the pepXML format. Most significant of these are full path exported for the data file, correct extension parsing, and putting ties inside the same search_hit element (with the same hit_rank=1). Also might have added support for MGF, but this is untested. Fixed a few compiler warnings.

+++ KIons.cpp
@@ -371,7 +371,7 @@ double KIons::getModMass(int index){
 }
 
 int KIons::getModMassSize(){
-  return modMassArray.size();
+  return (int)modMassArray.size();
 }
 
 double* KIons::getMods(){
